Add a new plugin for detection CORS Origin header reflection
The issue seems to be that some applications or servers will simply
reflected a given Origin header back in the Access-Control-Allow-Origin
response, leading to a CORS bypass. This also includes a case where
the reflected header includes a wildcard by fuzzy matching the header.

The issue is discussed further in:
http://blog.portswigger.net/2016/10/exploiting-cors-misconfigurations-for.html